Refactor : ChannelList의 ChannelListItem을 별도의 컴포넌트로 분리
