+++ backend/src/channel/channel.module.ts
@@ -8,7 +8,7 @@ import { Server } from '../server/server.entity';
 import { UserChannelService } from '../user-channel/user-channel.service';
 import { UserChannelRepository } from '../user-channel/user-channel.repository';
 import { UserRepository } from '../user/user.repository';
-import { ChannelRepository } from './user.repository';
+import { ChannelRepository } from './channel.repository';
 
 @Module({
   imports: [
