+++ frontend/src/components/Main/ChannelList.tsx
@@ -6,8 +6,7 @@ import { BoostCamMainIcons } from '../../utils/SvgIcons';
 import { ChannelData } from '../../types/main';
 import { MainStoreContext } from './MainStore';
 import ChannelListHeader from './ChannelListHeader';
-
-const { Hash } = BoostCamMainIcons;
+import ChannelListItem from './ChannelListItem';
 
 const Container = styled.div`
   width: 100%;
@@ -31,50 +30,11 @@ const ChannelListBody = styled.div`
   color: #a69c96;
   font-size: 15px;
 `;
-
-const ChannelListItem = styled.div<{ selected: boolean }>`
-  width: 100%;
-  height: 25px;
-
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-
-  box-sizing: border-box;
-  padding: 15px 0px 15px 25px;
-  ${(props) => (props.selected ? 'background-color:#21557C;' : '')}
-
-  &:hover {
-    cursor: pointer;
-    background-color: ${(props) => (props.selected ? '#21557C' : '#321832')};
-  }
-`;
-
-const ChannelNameSpan = styled.span`
-  padding: 5px 0px 5px 5px;
-`;
-
-const HashIcon = styled(Hash)`
-  width: 15px;
-  height: 15px;
-  fill: #a69c96;
-`;
-
 function ChannelList(): JSX.Element {
   const [isListOpen, setIsListOpen] = useState<boolean>(false);
-  const { selectedServer, selectedChannel, serverChannelList, setSelectedChannel } = useContext(MainStoreContext);
+  const [clickedChannel, setClickedChannel] = useState<number>(-1);
+  const { selectedServer, selectedChannel, serverChannelList } = useContext(MainStoreContext);
   const navigate = useNavigate();
-
-  const onClickChannelBlock = ({ currentTarget }: React.MouseEvent<HTMLDivElement>) => {
-    const channelId = currentTarget.dataset.id;
-    if (channelId) setSelectedChannel(channelId);
-  };
-
-  const onRightClickChannelItem = (e: React.MouseEvent<HTMLDivElement>) => {
-    e.preventDefault();
-  };
-
   useEffect(() => {
     const serverId = selectedServer?.server?.id || 'none';
 
@@ -88,18 +48,7 @@ function ChannelList(): JSX.Element {
 
   const listElements = serverChannelList.map((val: ChannelData): JSX.Element => {
     const selected = val.id === selectedChannel;
-    return (
-      <ChannelListItem
-        key={val.id}
-        data-id={val.id}
-        selected={selected}
-        onClick={onClickChannelBlock}
-        onContextMenu={onRightClickChannelItem}
-      >
-        <HashIcon />
-        <ChannelNameSpan>{val.name}</ChannelNameSpan>
-      </ChannelListItem>
-    );
+    return <ChannelListItem key={val.id} dataId={val.id} selected={selected} name={val.name} />;
   });
 
   return (
